Split up large vscode_sdfv.ts file
